port sofaboundaryconditions tests to sofang
