+++ Sofa/Component/Constraint/Projective/tests/ProjectDirectionConstraint_test.cpp
@@ -24,27 +24,28 @@ using sofa::testing::BaseSimulationTest;
 #include <sofa/testing/NumericTest.h>
 using sofa::testing::NumericTest;
 
-#include <SofaSimulationGraph/DAGSimulation.h>
+#include <sofa/simulation/graph/DAGSimulation.h>
 #include <sofa/defaulttype/VecTypes.h>
-#include <SofaBaseTopology/PointSetTopologyContainer.h>
-#include <SofaBoundaryCondition/ProjectToPlaneConstraint.h>
-#include <SofaBaseMechanics/MechanicalObject.h>
+#include <sofa/component/topology/container/dynamic/PointSetTopologyContainer.h>
+#include <sofa/component/constraint/projective/ProjectDirectionConstraint.h>
+#include <sofa/component/statecontainer/MechanicalObject.h>
 #include <sofa/core/MechanicalParams.h>
 #include <sofa/defaulttype/VecTypes.h>
 
 
-
 namespace sofa {
+
+namespace {
+
 using namespace component;
 using namespace defaulttype;
+using namespace core::objectmodel;
 
-
-
-/**  Test suite for ProjectToPlaneConstraint.
+/**  Test suite for ProjectDirectionConstraint.
 The test cases are defined in the #Test_Cases member group.
   */
 template <typename _DataTypes>
-struct ProjectToPlaneConstraint_test : public BaseSimulationTest, NumericTest<typename _DataTypes::Coord::value_type>
+struct ProjectDirectionConstraint_test : public BaseSimulationTest, NumericTest<typename _DataTypes::Coord::value_type>
 {
     typedef _DataTypes DataTypes;
     typedef typename DataTypes::VecCoord VecCoord;
@@ -53,48 +54,45 @@ struct ProjectToPlaneConstraint_test : public BaseSimulationTest, NumericTest<ty
     typedef typename DataTypes::Deriv Deriv;
     typedef typename DataTypes::CPos CPos;
     typedef typename Coord::value_type Real;
-    typedef projectiveconstraintset::ProjectToPlaneConstraint<DataTypes> ProjectToPlaneConstraint;
-    typedef typename ProjectToPlaneConstraint::Indices Indices;
-    typedef component::topology::PointSetTopologyContainer PointSetTopologyContainer;
-    typedef container::MechanicalObject<DataTypes> MechanicalObject;
+    typedef constraint::projective::ProjectDirectionConstraint<DataTypes> ProjectDirectionConstraint;
+    typedef typename ProjectDirectionConstraint::Indices Indices;
+    typedef component::topology::container::dynamic::PointSetTopologyContainer PointSetTopologyContainer;
+    typedef statecontainer::MechanicalObject<DataTypes> MechanicalObject;
 
     simulation::Node::SPtr root;                 ///< Root of the scene graph, created by the constructor an re-used in the tests
     simulation::Simulation* simulation;          ///< created by the constructor an re-used in the tests
 
     unsigned numNodes;                         ///< number of particles used for the test
     Indices indices;                           ///< indices of the nodes to project
-    CPos origin;                               ///< origin of the plane to project to
-    CPos normal;                               ///< normal of the plane to project to
-    typename ProjectToPlaneConstraint::SPtr projection;
+    CPos direction;                            ///< direction to project to
+    typename ProjectDirectionConstraint::SPtr projection;
     typename MechanicalObject::SPtr dofs;
 
-    /// Create the context for the matrix tests.
+     /// Create the context for the tests.
     void SetUp() override
     {
-//        if( sofa::simulation::getSimulation()==nullptr )
+        //Init
         sofa::simulation::setSimulation(simulation = new sofa::simulation::graph::DAGSimulation());
 
         /// Create the scene
         root = simulation->createNewGraph("root");
 
-        PointSetTopologyContainer::SPtr topology = core::objectmodel::New<PointSetTopologyContainer>();
+        PointSetTopologyContainer::SPtr topology = New<PointSetTopologyContainer>();
         root->addObject(topology);
 
-        dofs = core::objectmodel::New<MechanicalObject>();
+        dofs = New<MechanicalObject>();
         root->addObject(dofs);
 
-        projection = core::objectmodel::New<ProjectToPlaneConstraint>();
+        // Project direction constraint
+        projection = New<ProjectDirectionConstraint>();
         root->addObject(projection);
 
         /// Set the values
         numNodes = 3;
         dofs->resize(numNodes);
 
-
-        origin = CPos(0,0,0);
-        projection->f_origin.setValue(origin);
-        normal = CPos(1,1,1);
-        projection->f_normal.setValue(normal);
+        direction = CPos(0,1,1);
+        projection->f_direction.setValue(direction);
 
     }
 
@@ -102,21 +100,20 @@ struct ProjectToPlaneConstraint_test : public BaseSimulationTest, NumericTest<ty
       For each of these cases, we can test if the projections work
       */
     ///@{
-    /** Constrain one particle, and not the last one.
-    Detects bugs like not setting the projection matrix entries beyond the last constrained particle
+    /** Constraint one particle, and not the last one.
     */
     void init_oneConstrainedParticle()
     {
         indices.clear();
-        indices.push_back(1);
-        std::sort(indices.begin(),indices.end()); // checking vectors in linear time requires sorted indices
+        indices.push_back(0);
+//        std::sort(indices.begin(),indices.end()); // checking vectors in linear time requires sorted indices
         projection->f_indices.setValue(indices);
 
         /// Init
         sofa::simulation::getSimulation()->init(root.get());
     }
 
-    /** Constrain all the particles.
+    /** Constraint all the particles.
     */
     void init_allParticlesConstrained()
     {
@@ -130,7 +127,7 @@ struct ProjectToPlaneConstraint_test : public BaseSimulationTest, NumericTest<ty
     }
     ///@}
 
-
+    // Test project position
     bool test_projectPosition()
     {
        VecCoord xprev(numNodes);
@@ -138,7 +135,18 @@ struct ProjectToPlaneConstraint_test : public BaseSimulationTest, NumericTest<ty
        for (unsigned i=0; i<numNodes; i++){
            xprev[i] = x[i] = CPos(i,0,0);
        }
+
        projection->projectPosition(core::mechanicalparams::defaultInstance(), *dofs->write(core::VecCoordId::position()) );
+
+       type::vector<CPos> m_origin;
+
+       // particle original position
+       const VecCoord& xOrigin = projection->getMState()->read(core::ConstVecCoordId::position())->getValue();
+        for( typename Indices::const_iterator it = indices.begin() ; it != indices.end() ; ++it )
+        {
+             m_origin.push_back( DataTypes::getCPos(xOrigin[*it]) );
+        }
+
 
        bool succeed=true;
        typename Indices::const_iterator it = indices.begin(); // must be sorted
@@ -146,7 +154,9 @@ struct ProjectToPlaneConstraint_test : public BaseSimulationTest, NumericTest<ty
        {
            if ((it!=indices.end()) && ( i==*it ))  // constrained particle
            {
-              Real scal = (x[i]-origin)*normal; // null if x is in the plane
+              CPos crossprod = (x[i]-m_origin[i]).cross(direction); // should be parallel
+              Real scal = crossprod*crossprod; // null if x is on the line
+
               if( !this->isSmall(scal,100) ){
                   succeed = false;
                   ADD_FAILURE() << "Position of constrained particle " << i << " is wrong: " << x[i] ;
@@ -157,6 +167,7 @@ struct ProjectToPlaneConstraint_test : public BaseSimulationTest, NumericTest<ty
            {
               CPos dx = x[i]-xprev[i];
               Real scal = dx*dx;
+
               if( !this->isSmall(scal,100) ){
                   succeed = false;
                   ADD_FAILURE() << "Position of unconstrained particle " << i << " is wrong: " << x[i] ;
@@ -167,6 +178,7 @@ struct ProjectToPlaneConstraint_test : public BaseSimulationTest, NumericTest<ty
        return succeed;
     }
 
+    // Test project velocity
     bool test_projectVelocity()
     {
        VecDeriv vprev(numNodes);
@@ -174,15 +186,18 @@ struct ProjectToPlaneConstraint_test : public BaseSimulationTest, NumericTest<ty
        for (unsigned i=0; i<numNodes; i++){
            vprev[i] = v[i] = CPos(i,0,0);
        }
+
        projection->projectVelocity(core::mechanicalparams::defaultInstance(), *dofs->write(core::VecDerivId::velocity()) );
 
        bool succeed=true;
        typename Indices::const_iterator it = indices.begin(); // must be sorted
        for(unsigned i=0; i<numNodes; i++ )
        {
-           if ((it!=indices.end()) && ( i==*it ))  // constrained particle
+          if ((it!=indices.end()) && ( i==*it ))  // constrained particle
            {
-              Real scal = v[i]*normal; // null if v is in the plane
+              CPos crossprod = v[i].cross(direction); // should be parallel
+              Real scal = crossprod.norm(); // null if v is ok
+
               if( !this->isSmall(scal,100) ){
                   succeed = false;
                   ADD_FAILURE() << "Velocity of constrained particle " << i << " is wrong: " << v[i] ;
@@ -220,17 +235,18 @@ typedef Types<
 > DataTypes; // the types to instanciate.
 
 // Test suite for all the instanciations
-TYPED_TEST_SUITE(ProjectToPlaneConstraint_test, DataTypes);
+TYPED_TEST_SUITE(ProjectDirectionConstraint_test, DataTypes);
+
 // first test case
-TYPED_TEST( ProjectToPlaneConstraint_test , oneConstrainedParticle )
+TYPED_TEST( ProjectDirectionConstraint_test , oneConstrainedParticle )
 {
     EXPECT_MSG_NOEMIT(Error) ;
     this->init_oneConstrainedParticle();
     ASSERT_TRUE(  this->test_projectPosition() );
     ASSERT_TRUE(  this->test_projectVelocity() );
 }
-// next test case
-TYPED_TEST( ProjectToPlaneConstraint_test , allParticlesConstrained )
+// second test case
+TYPED_TEST( ProjectDirectionConstraint_test , allParticlesConstrained )
 {
     EXPECT_MSG_NOEMIT(Error) ;
     this->init_allParticlesConstrained();
@@ -238,7 +254,7 @@ TYPED_TEST( ProjectToPlaneConstraint_test , allParticlesConstrained )
     ASSERT_TRUE(  this->test_projectVelocity() );
 }
 
+} // anonymous namespace
 
 } // namespace sofa
-
