+++ Sofa/Component/Constraint/Projective/tests/ProjectToLineConstraint_test.cpp
@@ -24,28 +24,26 @@ using sofa::testing::BaseSimulationTest;
 #include <sofa/testing/NumericTest.h>
 using sofa::testing::NumericTest;
 
-#include <SofaSimulationGraph/DAGSimulation.h>
+#include <sofa/simulation/graph/DAGSimulation.h>
 #include <sofa/defaulttype/VecTypes.h>
-#include <SofaBaseTopology/PointSetTopologyContainer.h>
-#include <SofaBoundaryCondition/ProjectDirectionConstraint.h>
-#include <SofaBaseMechanics/MechanicalObject.h>
+#include <sofa/component/topology/container/dynamic/PointSetTopologyContainer.h>
+#include <sofa/component/constraint/projective/ProjectToLineConstraint.h>
+#include <sofa/component/statecontainer/MechanicalObject.h>
 #include <sofa/core/MechanicalParams.h>
 #include <sofa/defaulttype/VecTypes.h>
 
 
 namespace sofa {
-
-namespace {
-
 using namespace component;
 using namespace defaulttype;
-using namespace core::objectmodel;
 
-/**  Test suite for ProjectDirectionConstraint.
+
+
+/**  Test suite for ProjectToLineConstraint.
 The test cases are defined in the #Test_Cases member group.
   */
 template <typename _DataTypes>
-struct ProjectDirectionConstraint_test : public BaseSimulationTest, NumericTest<typename _DataTypes::Coord::value_type>
+struct ProjectToLineConstraint_test : public BaseSimulationTest, NumericTest<typename _DataTypes::Coord::value_type>
 {
     typedef _DataTypes DataTypes;
     typedef typename DataTypes::VecCoord VecCoord;
@@ -54,44 +52,47 @@ struct ProjectDirectionConstraint_test : public BaseSimulationTest, NumericTest<
     typedef typename DataTypes::Deriv Deriv;
     typedef typename DataTypes::CPos CPos;
     typedef typename Coord::value_type Real;
-    typedef projectiveconstraintset::ProjectDirectionConstraint<DataTypes> ProjectDirectionConstraint;
-    typedef typename ProjectDirectionConstraint::Indices Indices;
-    typedef component::topology::PointSetTopologyContainer PointSetTopologyContainer;
-    typedef container::MechanicalObject<DataTypes> MechanicalObject;
+    typedef constraint::projective::ProjectToLineConstraint<DataTypes> ProjectToLineConstraint;
+    typedef typename ProjectToLineConstraint::Indices Indices;
+    typedef component::topology::container::dynamic::PointSetTopologyContainer PointSetTopologyContainer;
+    typedef statecontainer::MechanicalObject<DataTypes> MechanicalObject;
 
     simulation::Node::SPtr root;                 ///< Root of the scene graph, created by the constructor an re-used in the tests
     simulation::Simulation* simulation;          ///< created by the constructor an re-used in the tests
 
     unsigned numNodes;                         ///< number of particles used for the test
     Indices indices;                           ///< indices of the nodes to project
-    CPos direction;                            ///< direction to project to
-    typename ProjectDirectionConstraint::SPtr projection;
+    CPos origin;                               ///< origin of the plane to project to
+    CPos direction;                            ///< direction of the line to project to
+    typename ProjectToLineConstraint::SPtr projection;
     typename MechanicalObject::SPtr dofs;
 
-     /// Create the context for the tests.
+    /// Create the context for the tests.
     void SetUp() override
     {
-        //Init
+//        if( sofa::simulation::getSimulation()==nullptr )
         sofa::simulation::setSimulation(simulation = new sofa::simulation::graph::DAGSimulation());
 
         /// Create the scene
         root = simulation->createNewGraph("root");
 
-        PointSetTopologyContainer::SPtr topology = New<PointSetTopologyContainer>();
+        PointSetTopologyContainer::SPtr topology = core::objectmodel::New<PointSetTopologyContainer>();
         root->addObject(topology);
 
-        dofs = New<MechanicalObject>();
+        dofs = core::objectmodel::New<MechanicalObject>();
         root->addObject(dofs);
 
-        // Project direction constraint
-        projection = New<ProjectDirectionConstraint>();
+        projection = core::objectmodel::New<ProjectToLineConstraint>();
         root->addObject(projection);
 
         /// Set the values
         numNodes = 3;
         dofs->resize(numNodes);
 
-        direction = CPos(0,1,1);
+
+        origin = CPos(0,0,0);
+        projection->f_origin.setValue(origin);
+        direction = CPos(1,1,1);
         projection->f_direction.setValue(direction);
 
     }
@@ -100,20 +101,21 @@ struct ProjectDirectionConstraint_test : public BaseSimulationTest, NumericTest<
       For each of these cases, we can test if the projections work
       */
     ///@{
-    /** Constraint one particle, and not the last one.
+    /** Constrain one particle, and not the last one.
+    Detects bugs like not setting the projection matrix entries beyond the last constrained particle
     */
     void init_oneConstrainedParticle()
     {
         indices.clear();
-        indices.push_back(0);
-//        std::sort(indices.begin(),indices.end()); // checking vectors in linear time requires sorted indices
+        indices.push_back(1);
+        std::sort(indices.begin(),indices.end()); // checking vectors in linear time requires sorted indices
         projection->f_indices.setValue(indices);
 
         /// Init
         sofa::simulation::getSimulation()->init(root.get());
     }
 
-    /** Constraint all the particles.
+    /** Constrain all the particles.
     */
     void init_allParticlesConstrained()
     {
@@ -127,7 +129,7 @@ struct ProjectDirectionConstraint_test : public BaseSimulationTest, NumericTest<
     }
     ///@}
 
-    // Test project position
+
     bool test_projectPosition()
     {
        VecCoord xprev(numNodes);
@@ -135,18 +137,7 @@ struct ProjectDirectionConstraint_test : public BaseSimulationTest, NumericTest<
        for (unsigned i=0; i<numNodes; i++){
            xprev[i] = x[i] = CPos(i,0,0);
        }
-
        projection->projectPosition(core::mechanicalparams::defaultInstance(), *dofs->write(core::VecCoordId::position()) );
-
-       type::vector<CPos> m_origin;
-
-       // particle original position
-       const VecCoord& xOrigin = projection->getMState()->read(core::ConstVecCoordId::position())->getValue();
-        for( typename Indices::const_iterator it = indices.begin() ; it != indices.end() ; ++it )
-        {
-             m_origin.push_back( DataTypes::getCPos(xOrigin[*it]) );
-        }
-
 
        bool succeed=true;
        typename Indices::const_iterator it = indices.begin(); // must be sorted
@@ -154,9 +145,8 @@ struct ProjectDirectionConstraint_test : public BaseSimulationTest, NumericTest<
        {
            if ((it!=indices.end()) && ( i==*it ))  // constrained particle
            {
-              CPos crossprod = (x[i]-m_origin[i]).cross(direction); // should be parallel
+              CPos crossprod = (x[i]-origin).cross(direction); // should be parallel
               Real scal = crossprod*crossprod; // null if x is on the line
-
               if( !this->isSmall(scal,100) ){
                   succeed = false;
                   ADD_FAILURE() << "Position of constrained particle " << i << " is wrong: " << x[i] ;
@@ -167,7 +157,6 @@ struct ProjectDirectionConstraint_test : public BaseSimulationTest, NumericTest<
            {
               CPos dx = x[i]-xprev[i];
               Real scal = dx*dx;
-
               if( !this->isSmall(scal,100) ){
                   succeed = false;
                   ADD_FAILURE() << "Position of unconstrained particle " << i << " is wrong: " << x[i] ;
@@ -178,7 +167,6 @@ struct ProjectDirectionConstraint_test : public BaseSimulationTest, NumericTest<
        return succeed;
     }
 
-    // Test project velocity
     bool test_projectVelocity()
     {
        VecDeriv vprev(numNodes);
@@ -186,7 +174,6 @@ struct ProjectDirectionConstraint_test : public BaseSimulationTest, NumericTest<
        for (unsigned i=0; i<numNodes; i++){
            vprev[i] = v[i] = CPos(i,0,0);
        }
-
        projection->projectVelocity(core::mechanicalparams::defaultInstance(), *dofs->write(core::VecDerivId::velocity()) );
 
        bool succeed=true;
@@ -197,7 +184,6 @@ struct ProjectDirectionConstraint_test : public BaseSimulationTest, NumericTest<
            {
               CPos crossprod = v[i].cross(direction); // should be parallel
               Real scal = crossprod.norm(); // null if v is ok
-
               if( !this->isSmall(scal,100) ){
                   succeed = false;
                   ADD_FAILURE() << "Velocity of constrained particle " << i << " is wrong: " << v[i] ;
@@ -235,18 +221,17 @@ typedef Types<
 > DataTypes; // the types to instanciate.
 
 // Test suite for all the instanciations
-TYPED_TEST_SUITE(ProjectDirectionConstraint_test, DataTypes);
-
+TYPED_TEST_SUITE(ProjectToLineConstraint_test, DataTypes);
 // first test case
-TYPED_TEST( ProjectDirectionConstraint_test , oneConstrainedParticle )
+TYPED_TEST( ProjectToLineConstraint_test , oneConstrainedParticle )
 {
     EXPECT_MSG_NOEMIT(Error) ;
     this->init_oneConstrainedParticle();
     ASSERT_TRUE(  this->test_projectPosition() );
     ASSERT_TRUE(  this->test_projectVelocity() );
 }
-// second test case
-TYPED_TEST( ProjectDirectionConstraint_test , allParticlesConstrained )
+// next test case
+TYPED_TEST( ProjectToLineConstraint_test , allParticlesConstrained )
 {
     EXPECT_MSG_NOEMIT(Error) ;
     this->init_allParticlesConstrained();
@@ -254,7 +239,6 @@ TYPED_TEST( ProjectDirectionConstraint_test , allParticlesConstrained )
     ASSERT_TRUE(  this->test_projectVelocity() );
 }
 
-} // anonymous namespace
 
 } // namespace sofa
