+++ Sofa/Component/MechanicalLoad/tests/scenes/TrianglePressureForceField.scn
