+++ Sofa/Component/Topology/Container/Grid/tests/CMakeLists.txt
@@ -4,7 +4,6 @@ project(Sofa.Component.Topology.Container.Grid_test)
 
 set(SOURCE_FILES
     CylinderGridTopology_test.cpp
-    RegularGridNodeCreation.h
     RegularGridTopology_test.cpp
     SparseGridTopology_test.cpp
     SphereGridTopology_test.cpp
