--- modules/SofaBoundaryCondition/SofaBoundaryCondition_test/scenes/scenes/TrianglePressureForceField.scn
@@ -1,12 +1,0 @@
-<?xml version="1.0"?>
-
-<Node 	name="Root" gravity="0 0 0" time="0" animate="0"  dt="0.5" showAxis="true">
-<MechanicalObject name="DOFs" showObject="1"  showObjectScale="5"  showIndices="1"  showIndicesScale="0.0003" position="0 0 0 1 0 0 0 1 0" />
-<MeshTopology name="triangle" triangles="0 1 2"  drawTriangles="1" position="@DOFs.position"/>
-
-<TriangleSetTopologyContainer name="TriangleContainer" triangles="@triangle.triangles"/>
-<TriangleSetTopologyModifier />
-<TriangleSetGeometryAlgorithms template="Vec3d" />
-<!--<TrianglePressureForceField normal="0 0 1" dmin="-0.01" dmax="0.01" showForces="1" pressure="0 0 0.6"/>-->
-
-</Node>
